Add CppUTest Unit Test Framework
tafthorne/fdev1 (the Debian equivalent of this image) includes the
CppUTest unit testing nad mocking framework for C/C++.  This image
should contain equivalent libraries.

There is an apt package available for Debian systems, so that is used to
install CppUTest on fdev1.  There is no package for CentOS so we must
install the code after compiling it ourselves.  We get the latest
release archive (v3.8 at present) and use that to compile and install
CppUTest before using Yum to install the rest of the libraries.

To keep the final image small and free of unnecessary tools we split the
build into a multi-stage build.  Call the first part builder as that is
what the stage does.  Install the GNU Autotools needed to compile
CppUtest.  Unpack the archive, make the code, check the code, install
the code and tidy up.

In the main (unlabeled) stage COPY the /usr/local/include/CppUTest
directory over from the builder stage for us to use.  Also copy the
other install artifacts over (.a and .pc files).