Add Protocol Buffers Library (from gRPC repo)
Add in the Protocol Buffers library based on the version included in the
gRPC repository.  The step after this will be to install the gRPC
libraries as well.

This addition was based on the instructions found at
https://github.com/grpc/grpc/blob/master/BUILDING.md
and the Dockerfile for the C++ gRPC image on Docker Hub found at
https://hub.docker.com/r/grpc/cxx/~/dockerfile/

I have ommitted a `make check` for the protobuf libraries at present as
this fails to execute on my Ubuntu machine and within my Docker Image
building test.

I tested that `protoc --help` executes within the Docker container.