Add gRPC Library
Add the gRPC library.

This addition was based on the instructions found at
https://github.com/grpc/grpc/blob/master/BUILDING.md
and the Dockerfile for the C++ gRPC image on Docker Hub found at
https://hub.docker.com/r/grpc/cxx/~/dockerfile/

which has been added as both the make system used to generate the
application and library code and the example makefiles use it to
determine the location of plugs and style of compiler.