Correct pkg-config And Libs Paths
CentOS and Debian (as used in fdev1) use differnt paths for their
standard filesystem layout.  Try to modify what we build and where it
goes to fit with the expected CentOS locations.

Add ARGuments to make updating the paths simpler in the future.

This image has been tested to see that it builds the gRPC examples and
can execute the route_guide one as well.